+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+[Ejercicio 2]
+- Definición de clase base EntidadBase
+- Definición de clase Genero con encapsulamiento
+- Definición de clase Editorial con encapsulamiento
+- Definición de clase Moneda con encapsulamiento
+- Definición de clase TipoCotizacion con encapsulamiento
+- Definición de clase Libro con relaciones a Genero y Editorial
+- Definición de clase Precio con relaciones a Libro y Moneda
+- Definición de clase Stock con relaciones a Libro
+- Definición de clase CotizacionDolar con relaciones a TipoCotizacion
+
 [Ejercicio 1]
 - Inicialización del repositorio en GitHub
 - Creación de la rama Sprint_1